Added new raster_steepest_descent module.
Added raster_steepest_descent module that holds functions that calculate
steepest descent on a raster grid. These functions are dynamically added
to RasterModelGrid as methods.
