+++ landlab/grid/raster.py
@@ -1954,208 +1954,6 @@ class RasterModelGrid(ModelGrid, RasterModelGridPlotter):
                             self._diag_activelink_tonode))
         )
 
-    def calculate_steepest_descent_across_cell_faces(self, *args, **kwds):
-        """rmg.calculate_steepest_descent_across_cell_faces(node_values, [cell_ids], return_node=False, out=None)
-        Steepest gradient to neighbor nodes.
-
-        Return the steepest downward slopes across cell faces, i.e., only to
-        the orthogonal nodes, ignoring any diagonals.
-
-        Calculate the gradients of *node_values*, given at every node in the
-        grid, relative to the nodes centered at *cell_ids*. Note that upward
-        slopes are reported as positive. That is, the gradient is positive if
-        a neighbor node's value is greater than that of the node as
-        *cell_ids*. Thus, the gradients reported by this method are likely
-        negative.
-
-        If *cell_ids* is not provided, calculate the steepest slope (most
-        negative gradient) for all cells in the grid.
-
-        Use the *out* keyword if you have an array that you want to put the
-        result into. If not given, create and return a new array.
-
-        Use the *return_node* keyword to also return the node id of the node
-        in the direction of the maximum gradient. In this case, the returned
-        object is a length 2 tuple of the two arrays, (gradients, node_IDs).
-
-        Parameters
-        ----------
-        node_values : array-like
-            Node quantities to take gradient of.
-        cell_ids : array-like, optional
-            Cell at which to calculate gradients.
-        return_node : boolean, optional
-            If ``True`` also return node ID to the steepest gradient.
-        out : array_like, optional
-            Alternative output array in which to place the result.  Must
-            be of the same shape and buffer length as the expected output.
-
-        See Also
-        --------
-        calculate_steepest_descent_across_adjacent_cells :
-            neighbors, including diagonals
-
-        Examples
-        --------
-        >>> from landlab import RasterModelGrid
-        >>> rmg = RasterModelGrid(3, 3)
-        >>> node_values = rmg.zeros()
-        >>> node_values[1] = -1
-        >>> rmg.calculate_steepest_descent_across_cell_faces(node_values, 0)
-        masked_array(data = [-1.],
-                     mask = False,
-               fill_value = 1e+20)
-        <BLANKLINE>
-
-        Get both the maximum gradient and the node to which the gradient is
-        measured.
-
-        >>> rmg.calculate_steepest_descent_across_cell_faces(node_values, 0,
-        ...     return_node=True)
-        (array([-1.]), array([1]))
-        """
-        return rfuncs.calculate_steepest_descent_across_cell_faces(
-            self, *args, **kwds)
-
-    def calculate_steepest_descent_across_cell_corners(self, *args, **kwds):
-        """rmg.calculate_steepest_descent_across_cell_corners(node_values [, cell_ids], return_node=False, out=None)
-        Steepest gradient to diagonal nodes.
-
-        Return the steepest descents, i.e., minimum gradients, across diagonal
-        cells.
-
-        Calculate the gradients of *node_values*, given at every node in the
-        grid, relative to the nodes centered at *cell_ids*. Note that upward
-        slopes are reported as positive. That is, the gradient is positive if
-        a neighbor node's value is greater than that of the node as *cell_ids*.
-
-        If *cell_ids* is not provided, calculate the minimum gradient for all
-        cells in the grid.
-
-        Use the *out* keyword if you have an array that you want to put the
-        result into. If not given, create and return a new array.
-
-        Use the *return_node* keyword to also return the node id of the node
-        in the direction of the maximum gradient. In this case, the returned
-        object is a length 2 tuple of the two arrays, (gradients, node_IDs).
-
-        Parameters
-        ----------
-        node_values : array-like
-            Node quantities to take gradient of.
-        cell_ids : array-like, optional
-            Cell at which to calculate gradients.
-        return_node : boolean, optional
-            If ``True`` also return node ID to the steepest gradient.
-        out : array_like, optional
-            Alternative output array in which to place the result.  Must
-            be of the same shape and buffer length as the expected output.
-
-        See Also
-        --------
-        calculate_steepest_descent_across_adjacent_cells :
-            neighbors, including diagonals
-        calculate_steepest_descent_across_cell_faces :
-            just neighbors
-
-        Examples
-        --------
-        >>> from landlab import RasterModelGrid
-        >>> rmg = RasterModelGrid(3, 3)
-        >>> node_values = rmg.zeros()
-        >>> node_values[0] = -1
-        >>> rmg.calculate_steepest_descent_across_cell_corners(node_values, 0)
-        array([-0.70710678])
-
-        Get both the maximum gradient and the node to which the gradient is
-        measured.
-
-        >>> rmg.calculate_steepest_descent_across_cell_corners(node_values, 0, return_node=True)
-        (array([-0.70710678]), array([0]))
-        """
-        return rfuncs.calculate_steepest_descent_across_cell_corners(
-            self, *args, **kwds)
-
-    def calculate_steepest_descent_across_adjacent_cells(self, node_values,
-                                                         *args, **kwds):
-        """rmg.calculate_steepest_descent_across_adjacent_cells(node_values, [cell_ids], method='d4', return_node=False, out=None)
-        Steepest gradient to adjoining nodes.
-
-        Calculate the steepest downward slopes, i.e., the most negative
-        gradients, of *node_values*, given at every node in the grid,
-        relative to the nodes centered at *cell_ids*. Return those (negative)
-        gradients. Note that upward slopes are reported as positive. That is,
-        the gradient is positive if a neighbor node's value is greater than
-        that of the node as *cell_ids*. This method handles both orthogonal
-        and diagonal neighbors.
-
-        If *cell_ids* is not provided, calculate the minimum gradient for all
-        cells in the grid. Note this is a cell ID, not a node ID.
-
-        The default is to only consider neighbor cells to the north, south,
-        east, and west. To also consider gradients to diagonal nodes, set the
-        *method* keyword to *d8* (the default is *d4*).
-
-        Use the *out* keyword if you have an array that you want to put the
-        result into. If not given, create a new array.
-
-        Use the *return_node* keyword to also the node id of the node in the
-        direction of the steepest slope. In this case, the returned object is
-        a length 2 tuple of the two arrays, (gradients, node_IDs).
-
-        Parameters
-        ----------
-        node_values : array-like
-            Node quantities to take gradient of.
-        cell_ids : array-like, optional
-            Cell at which to calculate gradients.
-        return_node : boolean, optional
-            If ``True`` also return node ID to the steepest gradient.
-        out : array_like, optional
-            Alternative output array in which to place the result.  Must
-            be of the same shape and buffer length as the expected output.
-
-        See Also
-        --------
-        calculate_steepest_descent_across_cell_corners :
-            just diagonals
-        calculate_steepest_descent_across_cell_faces :
-            just neighbors
-
-        Examples
-        --------
-        >>> from landlab import RasterModelGrid
-        >>> rmg = RasterModelGrid(4, 4)
-        >>> node_values = rmg.zeros()
-        >>> node_values[1] = -1
-        >>> rmg.calculate_steepest_descent_across_adjacent_cells(
-        ...     node_values, 0)
-        masked_array(data = [-1.],
-                     mask = False,
-               fill_value = 1e+20)
-        <BLANKLINE>
-
-        Get both the maximum gradient and the node to which the gradient is
-        measured.
-
-        >>> rmg.calculate_steepest_descent_across_adjacent_cells(
-        ...     node_values, 0, return_node=True)
-        (array([-1.]), array([1]))
-
-        Use method to choose which neighbors to consider.
-
-        >>> node_values[0] = -10.
-        >>> node_values[1] = -1.
-        >>> rmg.calculate_steepest_descent_across_adjacent_cells(
-        ...     node_values, 0, method='d4', return_node=True)
-        (array([-1.]), array([1]))
-        >>> rmg.calculate_steepest_descent_across_adjacent_cells(
-        ...     node_values, 0, method='d8', return_node=True)
-        (array([-7.07106781]), array([0]))
-        """
-        return rfuncs.calculate_steepest_descent_across_adjacent_cells(
-            self, node_values, *args, **kwds)
-
     def find_node_in_direction_of_max_slope(self, u, node_id):
         """Node of steepest gradient.
 
@@ -5007,4 +4805,6 @@ add_module_functions_to_class(RasterModelGrid, 'raster_mappers.py',
 add_module_functions_to_class(RasterModelGrid, 'raster_aspect.py',
                               pattern='calculate_slope_aspect*')
 add_module_functions_to_class(RasterModelGrid, 'raster_gradients.py',
+                              pattern='calculate_*')
+add_module_functions_to_class(RasterModelGrid, 'raster_steepest_descent.py',
                               pattern='calculate_*')
