+++ landlab/grid/tests/test_raster_funcs/test_gradients_across_cell_corners_max.py
@@ -7,8 +7,6 @@ try:
 except ImportError:
     from landlab.testing.tools import assert_is
 
-from landlab.grid import raster_funcs as rfuncs
-
 
 def setup_unit_grid():
     from landlab import RasterModelGrid
